hi
changed

+++ source.py
@@ -1,5 +1,7 @@
 
 def get_busyness_percentage(maximumCapacity, amountOfPeople):
-    return (amountOfPeople / maximumCapacity) * 100
+    return int(float(amountOfPeople) / float(maximumCapacity) * 100)
+
+
